requirements.txt
click==8.1.3
fernet==1.0.1
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pyaes==1.6.1
PyJWT==2.6.0
qrcode==7.3.1
Werkzeug==2.2.2
